+++ src/app-config/index.js
@@ -8,7 +8,6 @@ export * from '#x/ClassedConfig';
 export * from '#x/ConfigClassMapper';
 export * from '#x/EndpointConfig';
 export * from '#x/FileServiceConfig';
-export * from '#x/Files';
 export * from '#x/HostConfig';
 export * from '#x/MountConfig';
 export * from '#x/NamedConfig';